Move FunctionReturnValueNotUsedInspectionTests for QuickFixes into relevant files under RD.QuickFixes
